requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
pyTelegramBotAPI==3.7.6
python-dotenv==0.15.0
requests==2.25.1
urllib3==1.26.3
